pyproject.toml
[project]
name = "enrgdocdb"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask>=3.0.3",
    "flask-admin==2.0.0a2",
    "flask-security-too>=5.5.2",
    "setuptools>=75.2.0",
    "flask-sqlalchemy-lite>=0.1.0",
    "flask-alembic>=3.1.1",
    "argon2-cffi>=23.1.0",
    "flask-babel>=4.0.0",
    "requests>=2.32.3",
    "bootstrap-flask>=2.4.1",
    "cryptography>=43.0.3",
    "qrcode>=8.0",
    "python-dotenv>=1.0.1",
    "gunicorn>=23.0.0",
    "pymysql>=1.1.1",
    "pyjwt>=2.9.0",
    "authlib>=1.3.2",
    "mysqlclient>=2.2.6",
    "icalendar>=6.1.0",
    "flask-limiter>=3.10.1",
    "ldaptor>=21.2.0",
    "weasyprint>=60.1",
    "beautifulsoup4>=4.12.0",
    "bleach>=6.0.0",
]

[tool.uv]
dev-dependencies = [
    "djlint>=1.35.2",
    "factory-boy>=3.3.3",
    "pytest>=9.0.3",
    "pytest-flask>=1.3.0",
    "pytest-mock>=3.15.1",
    "ruff>=0.14.9",
]

[tool.djlint]
ignore = "H021"
